losetup and multipass
